Manage temp files (#30)
* :construction: Started to refractor purge to purge files
	[T] Keep refratoring each functions affected
	[#] Issues: #17
	[N] Some comments were deleted (cobra related)

* New func to display size + units

* :page_facing_up::toilet: Implemented the file purging
	[T] Tests
	[#] #30, #17

* :checkered_flag: Tests for purge
:point_right: Apply TreatLastChar() to conf.Gobin & conf.Gopath

+++ cmd/start.go
@@ -105,15 +105,23 @@ func callPurge(targets []string) error {
 		} */
 		/* pPath = path
 		purgeCmd.Execute() */
-		fInfo, errFInfo := fetchAll(path)
+		fInfo, single, errFInfo := fetchAll(path)
 		if errFInfo != nil {
 			color.HiRed("Error fetch info for this temp directory")
 			return errFInfo
 		}
-
-		if errPurgeAll := deleteAllStr(path, fInfo, testAll); errPurgeAll != nil {
-			color.HiRed("Error while purging all the file at once!")
-			return errPurgeAll
+		if single == nil {
+			// then it's a directory
+			if errPurgeAll := deleteAllStr(path, fInfo, testAll); errPurgeAll != nil {
+				color.HiRed("Error while purging directory:", path)
+				return errPurgeAll
+			}
+		} else {
+			// then it's a single file
+			if errSingleF := emptyFile(path, single, testAll); errSingleF != nil {
+				color.HiRed("Error while purging the single file:", path)
+				return errSingleF
+			}
 		}
 
 	}
